Se dividen los archivos en metodos que pueden ser importados en deferentes partes tambien se pasa toda la funcionalidad de la tabla a un archivo
