+++ resources/js/app.js
@@ -1,12 +1,19 @@
 require('./bootstrap');
-// importamos Swall
+/******* Importaciones *******/
+// Importamos Swal para usar las notificaciones
 import Swal from 'sweetalert2';
-import {  consultarContactos } from './consultarContactos.js'
+// se importa el metodo encargado del funcionamiento de la consuta de contactos
+import { consultarContactos } from './consultarContactos.js';
+import { 
+    limpiarErrores, 
+    notificarErrores, 
+    notificarErroresTelefono 
+} from './errores.js';
+import { validarForm } from './validarForm.js';
+import { notificacion } from './notificacion.js';
 /****** variables ******/
+// Este boton abre la modal principal para agregar contactos
 const btnAddContact = document.querySelector('#btnAddContact');
-const bodyTable = document.querySelector('#bodyTable');
-const telefonosContacto = document.querySelector('#telefonosContacto');
-const btnAddNumber = document.querySelector('#btnAddNumber');
 
 /****** Funciones ******/
 // agregar contacto
@@ -23,36 +30,12 @@ const agregarContacto = async () => {
     }else{
         // en caso de que no existan campos vacios realizamos la peticion
         const respuesta = await enviar( form );
-        console.log( 'respuesta', respuesta )
         if( respuesta.status == true ){
             notificacion( 'Bien...!', 'Registro correcto', 'success' );
             consultarContactos();
         }else{
             notificacion( 'Error...!', 'Uno o mas campos tienen error', 'error' );
             notificarErrores( respuesta.errores );
-        }
-    }
-}
-// agregar numero a usuario
-const agregarNumero = async () => {
-    const form = crearFormularioNumero();
-
-    limpiarErrores( form );
-    // reutilizamos el metodo para validar formulario
-    const validar = validarForm( form );
-    console.log( form );
-    // si la validacion es false notificamos el error
-    if( validar == false ){
-        notificacion( 'Error...!', 'Uno o mas campos estan vacios.....', 'error' );
-    }else{
-        // en caso de que no existan campos vacios realizamos la peticion
-        const respuesta = await enviarNumero( form );
-        console.log( 'respuesta', respuesta )
-        if( respuesta.status == true ){
-            notificacion( 'Bien...!', 'Registro correcto', 'success' );
-        }else{
-            notificacion( 'Error...!', 'Uno o mas campos tienen error', 'error' );
-            notificarErroresTelefono( respuesta.errores );
         }
     }
 }
@@ -71,50 +54,6 @@ const enviar = async ( form ) => {
         return err.response.data;
     }
 }
-// Metodo para enviar el formulario de numero
-const enviarNumero = async ( form ) => {
-    try{
-        const response = await axios.post("http://agenda-interesse.kame.house/api/phones", form);
-        const data = await response.data;
-        // limpiiamos el formulario y cerramos la modal
-        // document.getElementById("formAddContact").reset();
-        // document.getElementById('btnCloseModal').click();
-        return data;
-    }catch (err) {
-        // en caso de retornar un 422 de validacion del formulario
-        // console.log(err.response.data)
-        return err.response.data;
-    }
-}
-
-// metodo para notificar errores en el formulario
-const notificarErrores = ( errores ) => {
-    for (let clave in errores){
-        const elemento = `#${clave}`; 
-        const elementoError = `#${clave}-validate`;
-        document.querySelector(elemento).classList.add('is-invalid');
-        document.querySelector(elementoError).innerHTML =errores[clave][0];
-    }
-}
-// metodo para notificar errores en el formulario al agregar un telefono
-const notificarErroresTelefono = ( errores ) => {
-    for (let clave in errores){
-        const elemento = `#${clave}-add`; 
-        const elementoError = `#${clave}-validate-add`;
-        document.querySelector(elemento).classList.add('is-invalid');
-        document.querySelector(elementoError).innerHTML =errores[clave][0];
-    }
-}
-
-
-// metodo para limpiarErrores del fomulario 
-const limpiarErrores = ( form ) => {
-    for (let clave in form){
-        const elemento = `#${clave}`; 
-        document.querySelector(elemento).classList.remove('is-invalid');
-    }
-}
-
 // Metodo para crear la informacion del formulario
 const crearFormulario = () => {
     const form = {
@@ -128,36 +67,8 @@ const crearFormulario = () => {
     }
     return form;
 }
-// Metodo para crear la informacion del formulario para agregar numeros
-const crearFormularioNumero = () => {
-    const form = {
-        'alias_number':document.querySelector('#alias_number-add').value.trim(),
-        'number':document.querySelector('#number-add').value.trim(),
-        'user_id':document.querySelector('#user_id').value.trim()
-    }
-    return form;
-}
 
-// metodo de swall
-const notificacion = ( title, text, icon ) => {
-    Swal.fire({
-        title: title,
-        text: text,
-        icon: icon,
-    })
-}
 
-// Validamos que los ningun campo este vacio antes de enviar
-// El metodo retorna falso si encuentra un valor vacio 
-// Si no hay valores vacios retorna true.
-const validarForm = ( form ) => {
-    for (let clave in form){
-        if( form[clave] == "" ){
-            return false;
-        }
-    }
-    return true;
-}
 
 
 
@@ -165,7 +76,6 @@ const validarForm = ( form ) => {
 /****** AddEvents ******/
 const addEventsListener = () =>{
     btnAddContact.addEventListener("click", agregarContacto);
-    btnAddNumber.addEventListener("click", agregarNumero);
 }
 // Iniciamos los eventos
 addEventsListener();
